moved click echo formatting helpers to separate file
