+++ scripts/etherlink/deploy_erc20.py
@@ -2,6 +2,8 @@ import click
 from scripts.helpers.utility import (
     get_etherlink_web3,
     get_etherlink_account,
+)
+from scripts.helpers.formatting import (
     accent,
     echo_variable,
     wrap,
