+++ scripts/etherlink/withdraw.py
@@ -4,6 +4,8 @@ from scripts.helpers.utility import (
     get_etherlink_web3,
     get_etherlink_account,
     make_address_bytes,
+)
+from scripts.helpers.formatting import (
     accent,
     echo_variable,
     wrap,
