+++ scripts/helpers/utility.py
@@ -9,7 +9,6 @@ from pytezos.michelson.types.base import MichelsonType
 from typing import Any
 from web3 import Web3
 from eth_account.signers.local import LocalAccount
-import click
 
 
 # Default address used as a placeholder in the contract storage
@@ -128,15 +127,3 @@ def get_etherlink_account(web3: Web3, private_key: str) -> LocalAccount:
     # TODO: validate balance
 
     return account
-
-
-def accent(msg: str) -> str:
-    return click.style(msg, fg='bright_cyan')
-
-
-def wrap(msg: str, symbol: str = '`') -> str:
-    return '`' + msg + '`'
-
-
-def echo_variable(prefix: str, name: str, value: str) -> None:
-    click.echo(prefix + name + ': ' + wrap(accent(value)))
