+++ scripts/rollup_node/get_proof.py
@@ -1,7 +1,7 @@
 import click
 from scripts.helpers.rollup_node import get_proof as get_proof_from_rpc, Proof
 from scripts import cli_options
-from scripts.helpers.utility import accent
+from scripts.helpers.formatting import accent
 
 
 @click.command()
