+++ scripts/rollup_node/scan_outbox.py
@@ -1,7 +1,7 @@
 import click
 from scripts.helpers.rollup_node import get_messages
 from scripts import cli_options
-from scripts.helpers.utility import accent
+from scripts.helpers.formatting import accent
 import time
 import json
