+++ scripts/tezos/deploy_token.py
@@ -1,7 +1,7 @@
 import click
 from scripts.helpers.contracts import TokenHelper
-from scripts.helpers.utility import (
-    get_tezos_client,
+from scripts.helpers.utility import get_tezos_client
+from scripts.helpers.formatting import (
     accent,
     echo_variable,
     wrap,
